+++ tests/unit/server.spec.ts
@@ -1,5 +1,5 @@
 import { omit } from "ramda";
-import { makeRequestMock } from "../../src/mock";
+import { makeRequestMock } from "../../src/testing";
 import { givePort } from "../helpers";
 import { appMock, expressJsonMock, expressMock } from "../express-mock";
 import {
@@ -247,8 +247,11 @@ describe("Server", () => {
       expect(typeof rawPropMw).toBe("function");
       const buffer = Buffer.from([]);
       const requestMock = makeRequestMock({
-        method: "POST",
-        body: buffer,
+        mockFn: jest.fn,
+        requestProps: {
+          method: "POST",
+          body: buffer,
+        },
       });
       rawPropMw(requestMock, {}, jest.fn());
       expect(requestMock.body).toEqual({ raw: buffer });